Created Algorithm and EscapeAlgorithm parent classes to enable merging of A* & Dijkstra escape attempts (i.e. my and Dan's branches)
