+++ temple/src/main/java/student/Algorithm.java
@@ -1,0 +1,9 @@
+package student;
+
+import game.Node;
+
+import java.util.ArrayList;
+
+public abstract class Algorithm {
+    public abstract ArrayList<Node> shortestPath(Node _startNode, Node _endNode);
+}
